requirements.txt
Django>=4.2
django-unfold[all]>=0.15.0
django-import-export>=3.3.0
Pillow